requirements.txt
# Python Dependencies
# Document Intelligence Pipeline

# Web Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# OCR & Image Processing
pytesseract>=0.3.10
Pillow>=10.0.0
pdfplumber>=0.10.0
PyPDF2>=3.0.0

# Optional OCR Engine (uncomment to use)
# easyocr>=1.7.0
# torch>=2.0.0

# AI/ML
openai>=1.0.0

# Data Validation
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Utilities
pyyaml>=6.0.1
python-dotenv>=1.0.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0
httpx>=0.25.0

# Development (optional)
# black>=23.0.0
# flake8>=6.0.0
# mypy>=1.7.0
